move vars, helpers and assess funcs to fixtures.go
